back-end/requirements.txt

fastapi==0.95.2
openai>=1.68.2
python-dotenv==1.0.0
gitpython==3.1.32
pydantic==1.10.7
uvicorn==0.22.0